requirements.txt
PIL
numpy
PyOpenGL
PyOpenGL_accelerate
